+++ src/api/WalletUp.Infstructre/Services/ExchangeRateService.cs
@@ -6,7 +6,7 @@ namespace CashCat.Infstructre.Services;
 
 public class ExchangeRateService(IExchangeApi exchangeRateApi):IExchangeRateService
 {
-    public async Task<ExchangeApiResponse> GetRatesAsync(string currencies,decimal amount)
+    public async Task<ExchangeApiResponse> GetRatesAsync(string currencies,double amount)
     {
        var result= await exchangeRateApi.Convert(currencies,amount);
        return result;
